+++ src/main/java/com/brotes/api/modelo/itemPedidoRecurrente/DatosRespuestaItemRecurrente.java
@@ -3,7 +3,7 @@ package com.brotes.api.modelo.itemPedidoRecurrente;
 import com.brotes.api.modelo.categoria.Categoria;
 
 public record DatosRespuestaItemRecurrente(
-        int cantidad,
+        Double cantidad,
         Long id,
         String nombreProducto,
         Categoria categoria) {
